rename: 디렉토리명 오타 수정
